Renamed #todo and hooks for roller.
