+++ src/module/helpers/roll.js
@@ -35,7 +35,7 @@ export function proficiencyRollFormula(proficiency, attribute) {
 }
 
 export function proficiencyRoll(roller, proficiency, attribute) {
-  // TODO Likely need to add an actor activeEffectFormula to handle effects later.
+  // #todo Likely need to add an actor activeEffectFormula to handle effects later.
   let rollFormula = proficiencyRollFormula(proficiency, attribute);
   return roller(rollFormula);
 }
