+++ src/styles/farhome.less
@@ -1,4 +1,4 @@
-// TODO Make sure to prune unused CSS later
+// #todo Make sure to prune unused CSS later
 
 @import './components/_effects.less';
 @import './components/_forms.less';
