+++ src/styles/sheets/item-sheet.less
@@ -1,4 +1,4 @@
-// TODO Change this .sheet here to be specific to farhome item sheet
+// #todo Change this .sheet here to be specific to farhome item sheet
 
 .farhome.sheet.item {
   form {
